Added functions
I have added functions prior to numpy. It's a bit forced, as it
ought to really be part of the main language part, but I'd rather
have access to the concept of functions earlier rather than later.

Also, notebooks are fairly bad at good practice in coding (e.g.
make sure you have modular code and so on), so maybe this is a good
point to start with that mantra.