Added Linear model fitting with Mauna Loa session
Aims of this session are to familiarise the students with model
fitting, using initially very simple linear models, and to go through
a process of "eyeballing" it, building a brute force approach, and
then solving it using linear algebra (which also introduces linear
algebra in Python to them). In the process, the students should
build increasingly complex models of the Mauna Loa CO2 concentration
dataset, while at the same time critieque-ing them. Last bit is
using the models for prediction, which aims to kick off a discussion
on uncertaity, which doesn't fit here.