Fitting LAI from MODIS chapter added
I have added a chapter on fitting the MODIS LAI data. Aim is to
explore fitting and testing a non-linaer model with some data, as
well as exploring the uncertainty, which is done with MH sampling.
This is something they have already encountered briefly in the
methods course, and will be dealt with in terms of Bayesian inference
later on.